+++ src/views/library/components/library-items/list-items/book-list/bookitem-items/DeleteBtn.tsx
@@ -1,4 +1,4 @@
-import trashcan from '../../../../../assets/button/library/delete.png';
+import trashcan from '../../../../../../../assets/button/library/delete.png';
 
 const DeleteBtn = () => {
   return (
