+++ src/views/library/page/read-note/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ReadNotePage = () => {
+  return <div>기록페이지지</div>;
+};
+
+export default ReadNotePage;
